CLI framework [fixed #223]
https://github.com/turbinelabs/tbn/pull/233

I wanted to make a command-line application with many subcommands,
each with their own flags.

I started with [yacli](https://github.com/jonboulle/yacli/), which I
liked the behavior of but found difficult to reason about and test. In
particular, there was widespread use of global variables and init()
methods, and it felt a little soupy. I resolved to do whatever
minimally needed to be done to it to make it reasonably well factored
and easily tested. A week later, I have what amounts to a framework
for CLI, which I hope we can use for everything we write. It supports
global and per-subcommand flags, has built-in help and version
commands, and doesn't depend on weird init() calls or anything all
that peculiar.

Squashed commit of the following:

commit a8cc3a449b5b69d61b6c1be3c8300a492c273d70
Merge: 0982f10 5cba034
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Thu Feb 4 18:04:25 2016 -0800

    Merge branch 'master' into gds.cli#223

commit 0982f1069a3d7b217bc2c1ca0a76cd6c82a14c13
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Thu Feb 4 16:33:58 2016 -0800

    address comments

commit 8d3b1bd4747cee62f80e8213016fe775b74c47c7
Merge: 998a0f9 e8a152f
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Thu Feb 4 15:29:53 2016 -0800

    Merge branch 'master' into gds.cli#223

commit 998a0f94436e4aba2aa98780c67f7128736570c3
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Thu Feb 4 12:59:51 2016 -0800

    CLI framework

fbshipit-source-id: 584f956f3f668267e03ac2d0533abee49780f817