[cli] better usage output #405
https://github.com/turbinelabs/tbn/pull/405

This is an uglier diff than I was hoping for, BUT, this changes the
usage text to be friendlier and responsive to different terminal
widths. In addition to the layout, I made headers bold, and names of
subcommands and flags underlined. Most of the heavy lifting is done by
golang's doc package, with a shell out to `stty size` to get column
width

Squashed commit of the following:

commit 1ac43a362fae0321492120562567094913585467
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Wed Mar 23 19:57:37 2016 -0700

    don't use crypto

commit 54592b1f083d3b7c4bd368a0fe6fe8c71bc040a3
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Wed Mar 23 16:08:27 2016 -0700

    back out a bad merge

commit d1b18f77d05f96616716882d5a15a0eb0c8a8b95
Merge: a57ab58 1ce7d04
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Wed Mar 23 16:00:06 2016 -0700

    Merge branch 'master' into gds/cli_term

commit a57ab58b002198e5899262e5e9a834dcb74220e9
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Wed Mar 23 15:59:29 2016 -0700

    fnord

commit 026c3f9579bf573a414cad2b2e16a8280dbf10e3
Author: Glen D Sanford <glen@turbinelabs.io>
Date:   Wed Mar 23 15:56:01 2016 -0700

    Better command line stuff?

fbshipit-source-id: 998dcde