Introduce confagent, confagentctl [fixes #140] [fixes #141] (#522)
Yikes, I started this change in February.

This change introduces the `confagent` and `confagentctl` packages. Ignore any file prefixed with `mock_`.

## `confagent`
The confagent package introduces these interfaces:
- `Remote`: a thin layer on top of the API for (1) looking up a proxy by name and (2) pulling down the object graph for a proxy. I've just realized this should probably also include a ZoneKey. I'll fix that up, but wanted to get this review going.
- `ConfAgent`: polls for changes to a proxy's object graph, and writes out new nginx configs

There's an `nginxconfig` sub-package with an `NginxConfig` interface, which has a single `Write` method which is delegated to by the `ConfAgent`. Right now the only implementation is one that prints to STDOUT. We need to nail down the mechanism for HUPing NGINX before we go further here, but the plumbing is in place to configure an `NginxConfig` from flags.

For both `ConfAgent` and `NginxConfig` there are supporting `FromFlags` interfaces, which are used to configure a `flag.FlagSet` and then return a factory that will build the relevant objects based on the values in that `FlagSet`

### `confagentctl`

This is the command-line interface into `ConfAgent`. There are two commands:
- `poll`: fronts the `ConfAgent.Poll()` method. Need to add ZoneKey here too.
- `update`: reads `configwriter.ConfigObjects` off of STDIN, checks integrity, writes out to the configured `NginxConfig`.

### `configwriter`
- I added a FromFlags interface here to, to support configuring a `ConfigWriter` from flags.
- I made the fixtures accessable via `configwriter.MkObjectFixtures` function

### `assert/matcher`
I found it useful to have a `gomock.Matcher` that will fill an `io.Writer` with bytes. I added this package as a place to put useful generic Matchers as we develop them.

fbshipit-source-id: 4c06ab5