Required Flags for CLI [fixes #673] (#691)
I was trying to think of a reasonable way to centralize enforcement of required flags. It's cumbersome to hijack the various `FooVar` methods on FlagSet, it's similarly cumbersome to maintain/merge multiple FlagSets, and it's abhorrent to make our own. I also played around with adding a SetRequired method to command.Cmd and maintaining a []string of required flags.

In the end, I decided to hijack our existing convention of prefixing required flag usage with "[REQUIRED]". I added  `Required` and `IsRequired` methods to the `cli/flags` package, and use them in `Cmd.Run` to enforce requiredness. It's a little clumsy, but it's super easy to use when creating the flag, maps easily to the code in the field, and should survive copy-pasta.

The result is **a lot** less code, which seems good.

Also remove `confagentctl` which has been largely subsumed by `nginx-admin`.

fbshipit-source-id: b1d8e8c