[sdcollect] globalize updater flags (#876)
This makes more sense once #864 lands, but the idea is that all sdcollect backends will need an updater, so those flags should be global to sdcollect.

To make this be true, we need to be able to set the CLI's FlagSet, so that we can supply our own, wired to to the updater flags. Otherwise there's a circular dependency problem, since the commands are created before the CLI is.

So now we construct a FlagSet, wire it to an updater.FromFlags, then thread that to the Cmds, then push that FlagSet into the CLI.

The last part of the change is to move required-flag checking out of `Cmd` and into `CLI`, so that it can apply to global flags. It identifies all missing required flags, both global and per-command.

fbshipit-source-id: 4983ad3