library-ize dockerbuild & k8sconfiggen/deployment (#1314)
This lays groundwork for a fix for #1277 by exposing these tools via library. It's mostly moving things around, but I also introduced a `FromFlags` for `Deployment`, on the off chance we wanted to embed it in another CLI. For `Dockerbuild`, we will typically only want to build a name/tag, so the rest of the CLI flags aren't really that useful.

fbshipit-source-id: 668ccde