[cli/flag] facility for redacting sensitive flag values and warning about deprecated flag usage (fixes #2774) (fixes #2831) (#2818)
Currently, when we print CLI usage, we helpfully output environment variables that are being used to set flags. Something like:

```
    Options currently configured from the Environment:

        TBNCOLLECT_API_KEY=38763f72-a283-4ea7-888f-7f7a760c117d
```

This change does two things:

1. adds the ability to indicate that flags are sensitive, which is included as a prefix of the usage description, along with requiredness if appropriate:

```
    --api.key=string
            [REQUIRED/SENSITIVE] The auth key for API requests
```

2. Redacts these sensitive values from the environment variable output:

```
    Options currently configured from the Environment:

        TBNCOLLECT_API_KEY=<redacted>
```

To accomplish this, we introduce the `nonstdlib/flag/usage` package, which now houses the logic around marking flags as required or sensitive, and checking for those marks. Instead of looking for `"[REQUIRED] "` as a prefix, we now stuff some JSON into the usage string, which can be traded for a `Usage` interface, which has accessors and setters for the requiredness and sensitivity. `usage.New()` will eat either a regular string, or a json-encoded one, and in either case produce a sane `usage.Usage` interface. a `Pretty` method is provided to get the human-friendly output.

`flag.FromEnv` now redacts sensitive values in the response to the `Filled()` method, which takes care of the redaction. CLI calls `Pretty` when printing the flag usage, which takes care of the `"[REQUIRED/SENSITIVE]" prefixes.

In theory this Usage interface should let us add other annotations as we go, including custom validators.

For most callers `usage.Required()` and `usage.Sensitive()` should be sufficient, but when you want both, or optionality depending on circumstance, it's nicer to use `usage.New()`. The api key is a good example of that.

I also patched a few obvious places where api keys are being passed around. I'm sure I missed some, I mostly did it to prove out the api changes. I'll follow up with a change that replaces all calls to `tbnflag.[Is]Required` with calls to the `usage` package and removes those functions. I just didn't want to clutter up the diff more than I had to.

While I was in here, added support for marking flags deprecated too.

fbshipit-source-id: e470a7b