[cli] make it impossible to accidentally call command.Error() or command.BadInput() (#4120)
The CLI framework provides `Error[f]()` and `BadInput[f]()` methods on the `command.Cmd` struct to provide command-scoped errors. This is useful when you have a bunch of sub-commands, because it's obvious which command produced the error.

Until this change, it also provided no-scope equivalent functions, which caused confusion resulting in misuse. They were added so that the cli framework itself had a way to return errors before a command had been chosen, but actually only `command.BadInput` is called.

This change fixes all misuse, and removes these functions, and replaces the calls to `command.BadInput` in `cli.go` with a local private function.

fbshipit-source-id: 57734c6